Added notification_Screen and backGroundListener, it detects claps!
I created a backGroundListener class which extends service, this allows
the app to listen in the back ground, which it now does when the on
button is hit. It currently listens for a clap using a tarsus
percussiononset detector. If a clap is detected a notification_screen
activity will be triggered with the whatever the backGroundListener
tells it to be the message. Other then that cleaned up the screen
transitions with the intents using FLAG_ACTIVITY_NEW_TASK and
FLAG_ACTIVITY_CLEAR_TASK.