Revamp from todays meeting
The convolution is now straight out of wikkipedia, still using mfccs.
Now there is a control thread that calls listen() in bglistener it is
now more responsive. The convo values do now appear appear to have
significance to the sound around them, now we just need to determine
how to threshold. There are comments in the code, and we'll talk
tomorrow ;)